Obstacle tower placement (#21)
* Defense_Tower_Placement_Logic

Since the defense tower has four different prefabs for the four different positions relative to the enemy path, the tower building logic had to be updated to ensure the right prefab is used.

* Update_cost_for_defense_tower

The cost for a defense tower was reduced for development purposes.

* Refactoring by extraction

Moved the code for tower defense instantiation from the BuildTower method into a new method, that can be called by the BuildTower method. This ensures better readability and maintainability.