Boss Enemies (#34)
* Obstacle

adding an obstacle tower that only enemies that stops enemies that cant fly

* Damaging obstacle

enemies damage the obstacle and it gets destroyed after a while

* stealth enemy

script that makes the enemy invisible and undetectable by towers, until it is revealed

* Sniper tower script

Sniper tower can shoot invisible enemies

* Sniper script finish

Sniper Tower now reveals the stealth enemies and other towers can shoot them (animation and sprites still have to be added)

* scripts for new tower types

scripts and prefab structure created for the new towertypes and finished

* Fix bug

fix compiler error and error in script that occured afer pushing

* Defense Tower

adding animations and creating prefabs with the right sprites

* prefabs

create the prefabs with the sprites

* Obstacle tower placement (#21)

* Defense_Tower_Placement_Logic

Since the defense tower has four different prefabs for the four different positions relative to the enemy path, the tower building logic had to be updated to ensure the right prefab is used.

* Update_cost_for_defense_tower

The cost for a defense tower was reduced for development purposes.

* Refactoring by extraction

Moved the code for tower defense instantiation from the BuildTower method into a new method, that can be called by the BuildTower method. This ensures better readability and maintainability.

* Boss Enemy Script

Script implemented for the boss enemies. they spawn every second round and can destroy towers

* Sniper Tower Bug

Fixed the bug of the sniper Tower script where multiple projectiles are shot

* Boss Enemy Animation

Create the Animation for the Boss Enemy

* Tower Info

Add Info Screen for the description of the Tower Types

* Adjust scrollbar

adjust UI from the scrollbar

* Merge branch 'main' into Enemies

Merged the branch and fixed merge conflicts

* Enemy Spawner

change currentWave counter

* Sprites

fix sprites that got damaged during merging

* Chicken Projectile

* Shop Adjustements

adding the sniper and defense tower to the shop
adding help button to see tower descriptions

* fixed tower placing

---------

Co-authored-by: opourazar <104004211+opourazar@users.noreply.github.com>
Co-authored-by: Leonard <weasleywalker@gmx.de>